requirements.txt
# Core dependencies (existing)
openai>=1.3.0
anthropic>=0.21.0
python-dotenv>=1.0.0
pandas>=2.0.0
scipy>=1.11.0
streamlit>=1.28.0
plotly>=5.17.0

# NEW - Additional model providers
google-generativeai>=0.3.0
mistralai>=0.0.11
together>=1.0.0

# Optional: for better performance
requests>=2.31.0
numpy>=1.24.0